Rating Class, Assoc. Test
We decided in a meeting to modify our plan from the M2 to include Rating as it's own entity (so users can review w/o rating and vise versa, and so we can handle them separately). This is includes everything needed for Rating (service, model, schema) as well as a passing associated test file.